I've refactored `forwarding.py` into a modular package to improve its structure and maintainability. Here's a summary of the changes I've made:
- Replaced the monolithic `tgarchive/forwarding.py` with a new modular package structure in `tgarchive/forwarding/`.
- Broke down the forwarding logic into smaller, more manageable modules for client management, deduplication, message grouping, and queueing.
- Updated the tests to reflect the new modular structure.
